+++ module/Soutenance/src/Soutenance/Service/Notifier/NotifierSoutenanceService.php
@@ -71,7 +71,7 @@ class NotifierSoutenanceService extends NotifierService
      */
     protected function fetchEmailBdd(These $these) : string
     {
-        $variable = $this->variableService->getRepository()->findByCodeAndThese(Variable::CODE_EMAIL_BDD, $these);
+        $variable = $this->variableService->getRepository()->findOneByCodeAndThese(Variable::CODE_EMAIL_BDD, $these);
         return $variable->getValeur();
     }